requirements.txt
faiss-gpu-cu12==1.12.0
numpy>=1.26,<2
pypdf
rank-bm25
sentence-transformers
spacy
transformers
protobuf
sentencepiece
streamlit
openai
requests
